Move MemVisState into memvis.rs
