+++ src/io/memvis.rs
@@ -1,16 +1,37 @@
 //! Memory visualization
 
 use sdl2;
-use sdl2::rect::Point;
+use sdl2::rect::{Rect, Point};
 use sdl2::pixels::*;
-
-use io::constants::*;
-use cpu::constants::MemAddr;
-use cpu::*;
 
 use std::num::Wrapping;
 
+use io::constants::*;
+use io::graphics::Toggle;
+use cpu::constants::MemAddr;
+use cpu::*;
+
 use disasm;
+
+/// State for the memory visualization system
+pub struct MemVisState {
+    pub scale: f32,
+    pub tile_data_mode_button: Toggle<TileDataSelect>,
+    pub mem_val_display_enabled: bool,
+}
+
+impl MemVisState {
+    pub fn new() -> MemVisState {
+        MemVisState {
+            scale: SCALE,
+            tile_data_mode_button: Toggle::new(Rect::new(MEM_DISP_WIDTH, MEM_DISP_HEIGHT, 24, 12),
+                                               vec![TileDataSelect::Auto,
+                                                    TileDataSelect::Mode1,
+                                                    TileDataSelect::Mode2]),
+            mem_val_display_enabled: true,
+        }
+    }
+}
 
 
 /// Returns maybe a memory address given the coordinates of the memory visualization
